refactor: 뷰 분리 및 loadView를 사용해 분리한 뷰를 불러오도록 구현
