+++ Doesaegim/Doesaegim/Presentation/TravelScene/PlanAdd/View/PlanAddViewController.swift
@@ -5,8 +5,6 @@
 //  Created by 김민석 on 2022/11/16.
 //
 
-// swiftlint:disable file_length
-
 import UIKit
 
 import SnapKit
@@ -15,158 +13,7 @@ final class PlanAddViewController: UIViewController {
     
     // MARK: - UI properties
     
-    private lazy var scrollView: UIScrollView = {
-        let scrollView = UIScrollView()
-
-        scrollView.backgroundColor = .white
-        scrollView.showsVerticalScrollIndicator = false
-        return scrollView
-    }()
-    
-    private lazy var contentView: UIView = {
-        let view = UIView()
-        
-        return view
-    }()
-    
-    private lazy var planTitleStackView: UIStackView = {
-        let stackView = UIStackView()
-        
-        stackView.axis = .vertical
-        stackView.alignment = .fill
-        stackView.distribution = .fill
-        stackView.spacing = 12
-        return stackView
-    }()
-    
-    private lazy var planTitleLabel: UILabel = {
-        let label = UILabel()
-        
-        label.text = "일정 이름"
-        label.textColor = .black
-        label.font = .systemFont(ofSize: 20, weight: .bold)
-        label.textAlignment = .left
-        return label
-    }()
-    
-    private lazy var planTitleTextField: UITextField = {
-        let textField = UITextField()
-        
-        textField.placeholder = StringLiteral.planTextFieldPlaceHolder
-        textField.layer.cornerRadius = 10
-        textField.backgroundColor = .grey1
-        textField.textColor = .black
-        textField.font = .systemFont(ofSize: 17, weight: .regular)
-        textField.addPadding(witdh: 8)
-        textField.delegate = self
-        return textField
-    }()
-    
-    private lazy var placeTitleStackView: UIStackView = {
-        let stackView = UIStackView()
-        
-        stackView.axis = .vertical
-        stackView.alignment = .fill
-        stackView.distribution = .fill
-        stackView.spacing = 12
-        return stackView
-    }()
-    
-    private lazy var placeTitleLabel: UILabel = {
-        let label = UILabel()
-        
-        label.text = "장소"
-        label.textColor = .black
-        label.font = .systemFont(ofSize: 20, weight: .bold)
-        label.textAlignment = .left
-        return label
-    }()
-    
-    private lazy var placeSearchButton: UIButton = {
-        let button = UIButton()
-        
-        button.layer.cornerRadius = 10
-        button.backgroundColor = .grey1
-        button.setTitleColor(.grey3, for: .normal)
-        button.setTitle(StringLiteral.placeTextPlaceHolder, for: .normal)
-        button.titleLabel?.font = .systemFont(ofSize: 17, weight: .regular)
-        button.setImage(UIImage(systemName: "magnifyingglass"), for: .normal)
-        button.titleEdgeInsets = .init(top: 0, left: 15, bottom: 0, right: -5)
-        button.imageEdgeInsets = .init(top: 0, left: 5, bottom: 0, right: 5)
-        button.tintColor = .grey3
-        button.contentHorizontalAlignment = .left
-        return button
-    }()
-    
-    private lazy var dateStackView: UIStackView = {
-        let stackView = UIStackView()
-        
-        stackView.axis = .vertical
-        stackView.alignment = .fill
-        stackView.distribution = .fill
-        stackView.spacing = 12
-        return stackView
-    }()
-    
-    private lazy var dateTitleLabel: UILabel = {
-        let label = UILabel()
-        
-        label.text = "날짜와 시간"
-        label.textColor = .black
-        label.font = .systemFont(ofSize: 20, weight: .bold)
-        label.textAlignment = .left
-        return label
-    }()
-    
-    private lazy var dateInputButton: UIButton = {
-        let button = UIButton()
-        
-        button.layer.cornerRadius = 10
-        button.backgroundColor = .grey1
-        button.setTitleColor(.grey3, for: .normal)
-        button.setTitle(StringLiteral.datePlaceHolder, for: .normal)
-        button.titleLabel?.font = .systemFont(ofSize: 17, weight: .regular)
-        button.setImage(UIImage(systemName: "calendar"), for: .normal)
-        button.titleEdgeInsets = .init(top: 0, left: 15, bottom: 0, right: -5)
-        button.imageEdgeInsets = .init(top: 0, left: 5, bottom: 0, right: 5)
-        button.tintColor = .grey3
-        button.contentHorizontalAlignment = .left
-        return button
-    }()
-    
-    private lazy var descriptionTitleLabel: UILabel = {
-        let label = UILabel()
-        
-        label.text = "설명"
-        label.textColor = .black
-        label.font = .systemFont(ofSize: 20, weight: .bold)
-        label.textAlignment = .left
-        return label
-    }()
-    
-    private lazy var descriptionTextView: UITextView = {
-        let textView = UITextView()
-        
-        textView.layer.cornerRadius = 10
-        textView.text = StringLiteral.descriptionTextViewPlaceHolder
-        textView.textColor = .grey3
-        textView.font = .systemFont(ofSize: 14, weight: .regular)
-        textView.backgroundColor = .grey1
-        textView.contentInset = UIEdgeInsets(top: 8, left: 8, bottom: 8, right: 8)
-        textView.delegate = self
-        return textView
-    }()
-    
-    private lazy var addButton: UIButton = {
-        let button = UIButton()
-
-        button.setTitle("여행 추가", for: .normal)
-        button.setTitleColor(.white, for: .normal)
-        button.backgroundColor = .grey3
-        button.isEnabled = false
-        button.layer.cornerRadius = 10
-        return button
-    }()
+    private let rootView = PlanAddView()
     
     // MARK: - Properties
     
@@ -184,29 +31,16 @@ final class PlanAddViewController: UIViewController {
         fatalError("init(coder:) has not been implemented")
     }
     
+    override func loadView() {
+        view = rootView
+    }
+    
     override func viewDidLoad() {
         super.viewDidLoad()
-        configureViews()
-        dateInputButton.addTarget(
-            self,
-            action: #selector(dateInputButtonTouchUpInside),
-            for: .touchUpInside
-        )
-        planTitleTextField.addTarget(
-            self,
-            action: #selector(textFieldDidChange),
-            for: .editingChanged
-        )
-        placeSearchButton.addTarget(
-            self,
-            action: #selector(placeButtonTouchUpInside),
-            for: .touchUpInside
-        )
-        addButton.addTarget(
-            self,
-            action: #selector(addButtonTouchUpInside),
-            for: .touchUpInside)
-        viewModel.delegate = self
+        configureNavigationBar()
+        setKeyboardNotification()
+        setDelegate()
+        setAddTargets()
     }
     
     deinit {
@@ -214,82 +48,6 @@ final class PlanAddViewController: UIViewController {
     }
     
     // MARK: - Configure Functions
-    
-    private func configureViews() {
-        configureSubviews()
-        configureConstraint()
-        configureNavigationBar()
-        setKeyboardNotification()
-    }
-    
-    private func configureSubviews() {
-        view.addSubview(scrollView)
-        scrollView.addSubview(contentView)
-        contentView.addSubviews(
-            planTitleStackView, placeTitleStackView, dateStackView,
-            descriptionTitleLabel, descriptionTextView, addButton
-        )
-        planTitleStackView.addArrangedSubviews(planTitleLabel, planTitleTextField)
-        placeTitleStackView.addArrangedSubviews(placeTitleLabel, placeSearchButton)
-        dateStackView.addArrangedSubviews(dateTitleLabel, dateInputButton)
-        view.addGestureRecognizer(UITapGestureRecognizer(target: self, action: #selector(dismissKeyboard)))
-    }
-    
-    private func configureConstraint() {
-        scrollView.snp.makeConstraints {
-            $0.edges.equalToSuperview()
-        }
-        
-        contentView.snp.makeConstraints {
-            $0.edges.equalTo(scrollView.contentLayoutGuide)
-            $0.width.equalTo(scrollView.frameLayoutGuide)
-        }
-        
-        planTitleStackView.snp.makeConstraints {
-            $0.top.equalToSuperview().offset(17)
-            $0.leading.trailing.equalToSuperview().inset(16)
-        }
-        
-        planTitleTextField.snp.makeConstraints {
-            $0.height.equalTo(36)
-        }
-        
-        placeTitleStackView.snp.makeConstraints {
-            $0.top.equalTo(planTitleStackView.snp.bottom).offset(24)
-            $0.leading.trailing.equalToSuperview().inset(16)
-        }
-        
-        placeSearchButton.snp.makeConstraints {
-            $0.height.equalTo(36)
-        }
-        
-        dateStackView.snp.makeConstraints {
-            $0.top.equalTo(placeTitleStackView.snp.bottom).offset(24)
-            $0.leading.trailing.equalToSuperview().inset(16)
-        }
-        
-        dateInputButton.snp.makeConstraints {
-            $0.height.equalTo(36)
-        }
-        
-        descriptionTitleLabel.snp.makeConstraints {
-            $0.top.equalTo(dateStackView.snp.bottom).offset(24)
-            $0.leading.trailing.equalToSuperview().inset(16)
-        }
-        
-        descriptionTextView.snp.makeConstraints {
-            $0.top.equalTo(descriptionTitleLabel.snp.bottom).offset(12)
-            $0.leading.trailing.equalToSuperview().inset(16)
-            $0.height.equalTo(200)
-        }
-        
-        addButton.snp.makeConstraints {
-            $0.leading.trailing.equalToSuperview().inset(16)
-            $0.bottom.equalToSuperview().offset(-30)
-            $0.top.equalTo(descriptionTextView.snp.bottom).offset(40)
-            $0.height.equalTo(48)
-        }
-    }
     
     private func configureNavigationBar() {
         navigationItem.title = "일정 추가"
@@ -301,6 +59,37 @@ final class PlanAddViewController: UIViewController {
         )
     }
     
+    // MARK: - Set Delegate
+    
+    private func setDelegate() {
+        viewModel.delegate = self
+        rootView.planTitleTextField.delegate = self
+        rootView.descriptionTextView.delegate = self
+    }
+    
+    // MARK: - AddTarget
+    
+    private func setAddTargets() {
+        rootView.dateInputButton.addTarget(
+            self,
+            action: #selector(dateInputButtonTouchUpInside),
+            for: .touchUpInside
+        )
+        rootView.planTitleTextField.addTarget(
+            self,
+            action: #selector(textFieldDidChange),
+            for: .editingChanged
+        )
+        rootView.placeSearchButton.addTarget(
+            self,
+            action: #selector(placeButtonTouchUpInside),
+            for: .touchUpInside
+        )
+        rootView.addButton.addTarget(
+            self,
+            action: #selector(addButtonTouchUpInside),
+            for: .touchUpInside)
+    }
 }
 // MARK: - Keyboard
 
@@ -335,11 +124,11 @@ extension PlanAddViewController {
             bottom: keyboardSize.height - (tabBarController?.tabBar.frame.height ?? .zero),
             right: .zero
         )
-        scrollView.contentInset = contentInsets
+        rootView.scrollView.contentInset = contentInsets
     }
 
     @objc private func keyboardWillHide(notification: NSNotification) {
-        scrollView.contentInset = .zero
+        rootView.scrollView.contentInset = .zero
     }
     
     @objc private func dismissKeyboard() {
@@ -366,10 +155,10 @@ extension PlanAddViewController {
     
     @objc func addButtonTouchUpInside() {
         let result = viewModel.addPlan(
-            name: planTitleTextField.text,
-            dateString: dateInputButton.titleLabel?.text,
+            name: rootView.planTitleTextField.text,
+            dateString: rootView.dateInputButton.titleLabel?.text,
             locationDTO: locationDTO,
-            content: descriptionTextView.text
+            content: rootView.descriptionTextView.text
         )
         
         switch result {
@@ -383,10 +172,10 @@ extension PlanAddViewController {
     
     @objc func backButtonTouchUpInside() {
         viewModel.isClearInput(
-            title: planTitleTextField.text,
-            place: placeSearchButton.titleLabel?.text,
-            date: dateInputButton.titleLabel?.text,
-            description: descriptionTextView.text
+            title: rootView.planTitleTextField.text,
+            place: rootView.placeSearchButton.titleLabel?.text,
+            date: rootView.dateInputButton.titleLabel?.text,
+            description: rootView.descriptionTextView.text
         )
     }
 }
@@ -404,7 +193,7 @@ extension PlanAddViewController: UITextFieldDelegate {
 
 extension PlanAddViewController: UITextViewDelegate {
     func textViewDidBeginEditing(_ textView: UITextView) {
-        if textView.text == StringLiteral.descriptionTextViewPlaceHolder {
+        if textView.text == PlanAddView.StringLiteral.descriptionTextViewPlaceHolder {
             textView.text = nil
             textView.textColor = .black
         }
@@ -412,7 +201,7 @@ extension PlanAddViewController: UITextViewDelegate {
     
     func textViewDidEndEditing(_ textView: UITextView) {
         if textView.text.trimmingCharacters(in: .whitespacesAndNewlines).isEmpty {
-            textView.text = StringLiteral.descriptionTextViewPlaceHolder
+            textView.text = PlanAddView.StringLiteral.descriptionTextViewPlaceHolder
             textView.textColor = .grey3
         }
     }
@@ -422,12 +211,12 @@ extension PlanAddViewController: UITextViewDelegate {
 
 extension PlanAddViewController: PlanAddViewDelegate {
     func isVaildInputs(isValid: Bool) {
-        addButton.isEnabled = isValid
-        addButton.backgroundColor = isValid ? .primaryOrange : .grey3
+        rootView.addButton.isEnabled = isValid
+        rootView.addButton.backgroundColor = isValid ? .primaryOrange : .grey3
     }
     
     func planAddViewDidSelectLocation(locationName: String) {
-        placeSearchButton.setTitle(locationName, for: .normal)
+        rootView.placeSearchButton.setTitle(locationName, for: .normal)
     }
     
     func backButtonDidTap(isClear: Bool) {
@@ -443,11 +232,8 @@ extension PlanAddViewController: PlanAddViewDelegate {
             touchOption: .single, type: .dateAndTime, startDate: travel.startDate, endDate: travel.endDate
         )
         calendarViewController.delegate = self
-        
         present(calendarViewController, animated: true)
     }
-    
-    
 }
 
 // MARK: - SearchingLocationViewControllerDelegate
@@ -463,18 +249,7 @@ extension PlanAddViewController: SearchingLocationViewControllerDelegate {
 
 extension PlanAddViewController: CalendarViewDelegate {
     func fetchDate(dateString: String) {
-        dateInputButton.setTitle(dateString, for: .normal)
+        rootView.dateInputButton.setTitle(dateString, for: .normal)
         viewModel.isValidDate(dateString: dateString)
-    }
-}
-
-// MARK: - Namespaces
-
-extension PlanAddViewController {
-    enum StringLiteral {
-        static let planTextFieldPlaceHolder = "일정의 이름을 입력해주세요."
-        static let placeTextPlaceHolder = "장소를 검색해 주세요."
-        static let datePlaceHolder = "날짜와 시간을 입력해 주세요."
-        static let descriptionTextViewPlaceHolder = "설명을 입력해주세요."
     }
 }
